[FIX] res.groups: ensure predictable sort order
Unfortunately the group names are translatable,
and the display name is based on the hierarchy
of groups, so this will not give a satisfactory
result. What we would need ideally is to be able
to sort on the translated "full name" value,
but this is not supported by the ORM yet.
But a stable order is still better than no
order at all